Changed serialization of Datasets, and changed cache serialization to strings instead of binary code

+++ src/main.rs
@@ -28,18 +28,6 @@ pub fn index() -> HttpResponse {
     HttpResponse::Ok().body(html)
 }
 
-impl From<&dataset::Dataset> for sled::IVec {
-    fn from(dataset: &dataset::Dataset) -> sled::IVec {
-        bincode::serialize(dataset).unwrap().into()
-    }
-}
-
-impl From<sled::IVec> for dataset::Dataset {
-    fn from(ivec: sled::IVec) -> dataset::Dataset {
-        bincode::deserialize(&ivec).unwrap()
-    }
-}
-
 #[actix_rt::main]
 async fn main() -> std::io::Result<()> {
     dotenv().ok();
@@ -48,24 +36,12 @@ async fn main() -> std::io::Result<()> {
     let mut listenfd = ListenFd::from_env();
 
     let cache_path = env::var("CACHE_PATH").expect("Cache path not set");
-    std::fs::remove_dir_all(&cache_path).unwrap();
-    let t = sled::open(&cache_path).unwrap();
-    // let dataset1 = dataset::Dataset {
-    //     name: String::from("Dog photos"),
-    //     path: String::from("test_dataset"),
-    //     backend: backend::Backend::Local(backend::local::Local {
-    //         path: String::from("./storage/"),
-    //     }),
-    //     description: String::from("Very important dog photos."),
-    // };
-    // // t.insert(
-    // //     "test_dataset".as_bytes(),
-    // //     bincode::serialize(&dataset1).unwrap(),
-    // // )
-    // // .unwrap();
-    // t.insert("test_dataset", &dataset1).unwrap();
-    // let result: dataset::Dataset = t.get("test_dataset").unwrap().unwrap().into();
-    // debug!("result: {:?}", result);
+
+    // Delete the cache if it exists. (For development purposes.)
+    if std::path::Path::new(&cache_path).exists() {
+        std::fs::remove_dir_all(&cache_path).unwrap();
+    }
+    let t = sled::open(&cache_path).expect("Creation of cache db failed..");
 
     let config = config::Config {
         app_name: String::from("Actix-web"),
